Reorganise directory, improve comments and add more package managers
